requirements.txt
ollama
streamlit
pytest
faiss-cpu
sentence-transformers
numpy
python-docx
